Refactored all CUDA code to use Thrust
