+++ example/main.cpp
@@ -89,9 +89,6 @@ void profile()
 
 int main(int argc, char* argv[])
 {
-    profile();
-    return 0;
-
     NuclearReactor reactor = nuclear_reactors::OSIRIS();
     Material material = materials::SA304();
