+++ include/cluster_dynamics_state.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <vector>
+
+struct ClusterDynamicsState
+{
+    bool valid = true;
+    double time = 0.0;
+    std::vector<double> interstitials;
+    std::vector<double> vacancies;
+    double dislocation_density = 0.0;
+};
